# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-market investment portfolio tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portfolios:
    get:
      operationId: listPortfolios
      tags: [portfolios]
      summary: List all portfolios for the current user
      responses:
        "200":
          description: List of portfolios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Portfolio"
    post:
      operationId: createPortfolio
      tags: [portfolios]
      summary: Create a new portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
  /portfolios/{portfolioId}:
    get:
      operationId: getPortfolio
      tags: [portfolios]
      summary: Get a single portfolio
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "404":
          description: Not found
    patch:
      operationId: updatePortfolio
      tags: [portfolios]
      summary: Update a portfolio
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
    delete:
      operationId: deletePortfolio
      tags: [portfolios]
      summary: Delete a portfolio
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /portfolios/{portfolioId}/summary:
    get:
      operationId: getPortfolioSummary
      tags: [portfolios]
      summary: Get portfolio summary (current value, gain/loss, cash, dividends)
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Portfolio summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioSummary"
  /portfolios/{portfolioId}/performance:
    get:
      operationId: getPortfolioPerformance
      tags: [portfolios]
      summary: Get historical performance data points for charting
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
            enum: [1M, 3M, 6M, 1Y, ALL]
            default: 1Y
      responses:
        "200":
          description: Performance data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerformancePoint"

  # ── HOLDINGS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Portfolio:
      type: object
      required: [id, userId, name, baseCurrency, type, defaultFeeRate, sellFeeRate, defaultTaxRate, cashBalance, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        baseCurrency:
          type: string
          description: ISO currency code e.g. PHP, USD, GBP
        type:
          type: string
          enum: [PSE, LSE, US, CRYPTO, CUSTOM, BONDS, FUNDS, MIXED]
        defaultFeeRate:
          type: number
          description: Buy fee rate as percentage (e.g. 0.295 means 0.295%)
        sellFeeRate:
          type: number
          description: Sell fee and tax rate as percentage (e.g. 0.295 means 0.295%)
        defaultTaxRate:
          type: number
          description: Dividend withholding tax rate as percentage (e.g. 10 means 10%)
        cashBalance:
          type: number
        notes:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
          description: Hex color code for this portfolio (e.g. "#3b82f6")
        createdAt:
          type: string
    PortfolioInput:
      type: object
      required: [name, baseCurrency, type]
      properties:
        name:
          type: string
          minLength: 1
        baseCurrency:
          type: string
        type:
          type: string
          enum: [PSE, LSE, US, CRYPTO, CUSTOM, BONDS, FUNDS, MIXED]
        defaultFeeRate:
          type: number
          default: 0
        sellFeeRate:
          type: number
          default: 0
        defaultTaxRate:
          type: number
          default: 0
        notes:
          type: string
        color:
          type: ["string", "null"]
    PortfolioUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        baseCurrency:
          type: string
        type:
          type: string
          enum: [PSE, LSE, US, CRYPTO, CUSTOM, BONDS, FUNDS, MIXED]
        defaultFeeRate:
          type: number
        sellFeeRate:
          type: number
        defaultTaxRate:
          type: number
        notes:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
    PortfolioSummary:
      type: object
      required: [portfolioId, totalValue, costBasis, unrealizedGain, unrealizedGainPct, cashBalance, totalDividends, holdingCount]
      properties:
        portfolioId:
          type: integer
        totalValue:
          type: number
        costBasis:
          type: number
        unrealizedGain:
          type: number
        unrealizedGainPct:
          type: number
        cashBalance:
          type: number
        totalDividends:
          type: number
        holdingCount:
          type: integer
        monthlyDividends:
          type: number
        netDeposited:
          type: number
          description: Sum of DEPOSIT transactions minus sum of WITHDRAWAL transactions
        realizedGain:
          type: number
          description: Net gain from completed SELL transactions
        capitalGains:
          type: number
          description: Unrealized gain + realized gain
        dividendIncome:
          type: number
          description: Sum of all DIVIDEND type transactions
        cashGains:
          type: number
          description: Sum of all CASH_GAIN type transactions
        totalGains:
          type: number
          description: Capital gains + dividend income + cash gains
        totalGainsPct:
          type: number
          description: Total gains as percentage of net deposited
    PerformancePoint:
      type: object
      required: [date, value]
      properties:
        date:
          type: string
        value:
          type: number
        cashBalance:
          type: number
...